requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
ollama
langchain
langchain-community
pyyaml
